Improved Travis logging.

+++ travis.sh
@@ -41,10 +41,11 @@ trap 'error_handler' ERR
 bash -c "while true; do echo \$(date) - building ...; sleep $PING_SLEEP; done" &
 PING_LOOP_PID=$!
 
-# My build is using maven, but you could build anything with this, E.g.
-# your_build_command_1 >> $BUILD_OUTPUT 2>&1
-# your_build_command_2 >> $BUILD_OUTPUT 2>&1
-mvn install -PwithRatCheck >> $BUILD_OUTPUT 2>&1
+if [ -e "${DEPLOY}" ]; then
+    mvn clean deploy -DskipTests --settings ~/.m2/mySettings.xml >> $BUILD_OUTPUT 2>&1
+else
+    mvn install -PwithRatCheck >> $BUILD_OUTPUT 2>&1
+fi
 
 # The build finished without returning an error so dump a tail of the output
 dump_output
